Dao and service creation

+++ src/main/java/entities/Year.java
@@ -5,6 +5,7 @@
 package entities;
 
 import jakarta.persistence.*;
+
 import java.util.Objects;
 
 /**
@@ -18,7 +19,7 @@ public class Year extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Long id;
+    private Integer id;
 
     @Column(name = "value", nullable = false)
     private Integer value;
@@ -26,11 +27,11 @@ public class Year extends BaseEntity {
     public Year() {
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
